Got rid of all the Mark* methods and replaced them with Mark which returns a TypeMarker that has methods for each Marking. (Mark<SomeType>().Stubbed())
Made it so the Get methods return mocks even if a Create hasn't happened yet for those mocks. This allows you to prefetch mocks and set them up before a Create.

Made it so you have to give the AutoMockingContainer a MockRepository in its constructor rather than each of its methods. No reason to be able to change that on the fly.

git-svn-id: https://rhino-tools.svn.sourceforge.net/svnroot/rhino-tools/trunk@689 079b0acf-d9fa-0310-9935-e5ade295c882