More than significantly improving the performance of Rhino.Mocks when mocking many things in a single run by working around DynamicProxy2's performance limitations.
This may be a bug in DP2, but for now this gets around it and doesn't seem to have any ill effects. 

Basically, instead of just having a single ProxyGenerator, which I assume causes a single in memory assembly to continue to grow (each additional growth causes a linear increase in proxy generation time), we keep a map of Types to ProxyGenerators so we still get DP's caching benefits but none of its detriments.

git-svn-id: https://rhino-tools.svn.sourceforge.net/svnroot/rhino-tools/trunk@1382 079b0acf-d9fa-0310-9935-e5ade295c882